Fix render parts structure baseline
